Separation of bubble bounce component
